# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TradeDesk-AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/cocaptain/cards/{pending_id}/approve": {
      "post": {
        "summary": "Approve",
        "operationId": "approve_api_cocaptain_cards__pending_id__approve_post",
        "parameters": [
          {
            "name": "pending_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pending Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResult"
                }
              }
            }
          },
          "404": {
            "description": "Not a card you may review"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalConflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cocaptain/cards/{pending_id}/decline": {
      "post": {
        "summary": "Decline",
        "operationId": "decline_api_cocaptain_cards__pending_id__decline_post",
        "parameters": [
          {
            "name": "pending_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pending Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingOrder"
                }
              }
            }
          },
          "404": {
            "description": "Not a card you may review"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalConflict"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalConflict": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "HASH_MISMATCH",
              "EXPIRED",
              "NOT_PENDING",
              "REQUOTE_REQUIRED",
              "BLOCKED",
              "ACK_REQUIRED",
              "AWAITING_CO_CAPTAIN"
            ],
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "pending": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingOrder"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Plan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "pending",
          "plan"
        ],
        "title": "ApprovalConflict",
        "description": "Body of a 409 from the approve route. Nothing was sent."
      },
...
      "Charges": {
        "properties": {
          "brokerage": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Brokerage",
            "default": 0
          },
          "stt": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stt",
            "default": 0
          },
          "exchange_txn": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Exchange Txn",
            "default": 0
          },
          "sebi_fee": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sebi Fee",
            "default": 0
          },
          "stamp_duty": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stamp Duty",
            "default": 0
          },
          "gst": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Gst",
            "default": 0
          },
          "clearing": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Clearing",
            "default": 0
          },
          "ipft": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Ipft",
            "default": 0
          },
          "dp_charge": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Dp Charge",
            "default": 0
          },
          "break_even_price": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Break Even Price",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "brokerage",
          "stt",
          "exchange_txn",
          "sebi_fee",
          "stamp_duty",
          "gst",
          "clearing",
          "ipft",
          "dp_charge",
          "break_even_price",
          "total"
        ],
        "title": "Charges",
        "description": "021's charge breakdown for one order, in paise.\n\n`clearing` is kept for the plan's 8-part list; 021's pricing page lists no clearing\ncharge for equity, so it is 0. `dp_charge` is the per-ISIN depository charge 021\nlevies on delivery sells."
      },
...
      "Exchange": {
        "type": "string",
        "enum": [
          "NSE",
          "BSE"
        ],
        "title": "Exchange"
      },
      "ExecutionResult": {
        "properties": {
          "pending": {
            "$ref": "#/components/schemas/PendingOrder"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "SENT",
              "REJECTED",
              "UNKNOWN"
            ],
            "title": "Outcome"
          },
          "order": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Order"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pending",
          "outcome",
          "order",
          "message"
        ],
        "title": "ExecutionResult",
        "description": "What happened after an approved card was sent.\n\nSENT      the broker accepted it (see `order` for its live status)\nREJECTED  the broker refused it (`order.rejection_reason` says why)\nUNKNOWN   the request timed out and we could not confirm either way. The order is\n          never re-sent; it is reconciled against the broker's order book."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Instrument": {
        "properties": {
          "symbol": {
            "type": "string",
            "pattern": "^[A-Z0-9&\\-_.]{1,40}$",
            "title": "Symbol"
          },
          "exchange": {
            "$ref": "#/components/schemas/Exchange"
          },
          "series": {
            "type": "string",
            "title": "Series",
            "default": "EQ"
          },
          "isin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isin"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "tick_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Tick Size",
            "default": 5
          },
          "price_band_low": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Band Low"
          },
          "price_band_high": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Band High"
          },
          "suspended": {
            "type": "boolean",
            "title": "Suspended",
            "default": false
          },
          "underlying": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Underlying"
          },
          "lot_size": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Size"
          },
          "expiry": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry"
          },
          "strike": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike"
          },
          "option_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptionType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "exchange",
          "series",
          "isin",
          "name",
          "tick_size",
          "price_band_low",
          "price_band_high",
          "suspended",
          "underlying",
          "lot_size",
          "expiry",
          "strike",
          "option_type"
        ],
        "title": "Instrument"
      },
...
      "LegFailurePolicy": {
        "type": "string",
        "enum": [
          "HALT",
          "CONTINUE"
        ],
        "title": "LegFailurePolicy"
      },
...
      "OptionType": {
        "type": "string",
        "enum": [
          "CE",
          "PE"
        ],
        "title": "OptionType"
      },
      "Order": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "instrument": {
            "$ref": "#/components/schemas/Instrument"
          },
          "side": {
            "$ref": "#/components/schemas/Side"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          },
          "filled_quantity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Filled Quantity",
            "default": 0
          },
          "avg_fill_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Fill Price"
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderType"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "trigger_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Price"
          },
          "product": {
            "$ref": "#/components/schemas/Product",
            "default": "CNC"
          },
          "validity": {
            "$ref": "#/components/schemas/Validity",
            "default": "DAY"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RejectionReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "rejection_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "pending_quantity": {
            "type": "integer",
            "title": "Pending Quantity",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "order_id",
          "instrument",
          "side",
          "quantity",
          "filled_quantity",
          "avg_fill_price",
          "order_type",
          "limit_price",
          "trigger_price",
          "product",
          "validity",
          "status",
          "rejection_reason",
          "rejection_message",
          "created_at",
          "updated_at",
          "pending_quantity"
        ],
        "title": "Order",
        "description": "Broker-side order state, as reported by the broker."
      },
      "OrderAction": {
        "type": "string",
        "enum": [
          "PLACE",
          "MODIFY",
          "CANCEL"
        ],
        "title": "OrderAction"
      },
...
      "OrderStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "OPEN",
          "PARTIAL",
          "FILLED",
          "REJECTED",
          "CANCELLED",
          "UNKNOWN"
        ],
        "title": "OrderStatus"
      },
      "OrderType": {
        "type": "string",
        "enum": [
          "LIMIT",
          "MARKET",
          "STOP_LIMIT"
        ],
        "title": "OrderType"
      },
...
      "PendingOrder": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "action": {
            "$ref": "#/components/schemas/OrderAction"
          },
          "instrument": {
            "$ref": "#/components/schemas/Instrument"
          },
          "side": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Side"
              },
              {
                "type": "null"
              }
            ]
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "order_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderType"
              },
              {
                "type": "null"
              }
            ]
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "protection_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Protection Price"
          },
          "trigger_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Price"
          },
          "product": {
            "$ref": "#/components/schemas/Product",
            "default": "CNC"
          },
          "validity": {
            "$ref": "#/components/schemas/Validity",
            "default": "DAY"
          },
          "target_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Order Id"
          },
          "client_order_id": {
            "type": "string",
            "title": "Client Order Id"
          },
          "charges": {
            "$ref": "#/components/schemas/Charges"
          },
          "est_total": {
            "type": "integer",
            "title": "Est Total",
            "default": 0
          },
          "ref_ltp": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Ref Ltp"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "state": {
            "$ref": "#/components/schemas/PendingState",
            "default": "PENDING"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "risk_ack_required": {
            "type": "boolean",
            "title": "Risk Ack Required",
            "default": false
          },
          "co_captain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Co Captain"
          },
          "co_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Co Reasons"
          },
          "order_hash": {
            "type": "string",
            "title": "Order Hash",
            "description": "Hash of the exact order. Excludes LTP, expiry and state on purpose.\n\nPrice drift is handled by the drift check and re-quote; the hash only binds what\nthe trader approved: what, how much, at what price, under which idempotency key.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "action",
          "instrument",
          "side",
          "quantity",
          "order_type",
          "limit_price",
          "protection_price",
          "trigger_price",
          "product",
          "validity",
          "target_order_id",
          "client_order_id",
          "charges",
          "est_total",
          "ref_ltp",
          "created_at",
          "expires_at",
          "state",
          "plan_id",
          "rule_id",
          "warnings",
          "risk_ack_required",
          "co_captain",
          "co_reasons",
          "order_hash"
        ],
        "title": "PendingOrder",
        "description": "An exact order awaiting the trader's click. Built by code only."
      },
...
      "PendingState": {
        "type": "string",
        "enum": [
          "PENDING",
          "AWAITING_CO_APPROVAL",
          "APPROVED",
          "SENT",
          "REQUOTE_REQUIRED",
          "EXPIRED",
          "VOID",
          "REJECTED"
        ],
        "title": "PendingState"
      },
...
      "Plan": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/PlanLeg"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Legs"
          },
          "on_leg_failure": {
            "$ref": "#/components/schemas/LegFailurePolicy",
            "default": "HALT"
          },
          "state": {
            "$ref": "#/components/schemas/PlanState",
            "default": "PENDING"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "plan_hash": {
            "type": "string",
            "title": "Plan Hash",
            "description": "Binds the approval to every leg, its caps, and the failure policy.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "legs",
          "on_leg_failure",
          "state",
          "created_at",
          "expires_at",
          "plan_hash"
        ],
        "title": "Plan",
        "description": "Several orders approved together (shown to the trader as a basket order)."
      },
...
      "PlanLeg": {
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Index"
          },
          "order": {
            "$ref": "#/components/schemas/PendingOrder"
          },
          "quantity_basis": {
            "$ref": "#/components/schemas/QuantityBasis",
            "default": "FIXED"
          },
          "proceeds_from_leg": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proceeds From Leg"
          },
          "max_quantity": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Quantity"
          },
          "max_spend": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Spend"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "index",
          "order",
          "quantity_basis",
          "proceeds_from_leg",
          "max_quantity",
          "max_spend"
        ],
        "title": "PlanLeg"
      },
...
      "PlanState": {
        "type": "string",
        "enum": [
          "PENDING",
          "REQUOTE_REQUIRED",
          "APPROVED",
          "RUNNING",
          "COMPLETED",
          "HALTED",
          "EXPIRED",
          "VOID",
          "REJECTED"
        ],
        "title": "PlanState"
      },
...
      "Product": {
        "type": "string",
        "enum": [
          "CNC",
          "MIS",
          "NRML"
        ],
        "title": "Product"
      },
...
      "QuantityBasis": {
        "type": "string",
        "enum": [
          "FIXED",
          "FROM_PROCEEDS"
        ],
        "title": "QuantityBasis"
      },
...
      "RejectionReason": {
        "type": "string",
        "enum": [
          "INSUFFICIENT_FUNDS",
          "PRICE_BAND",
          "INVALID_PRICE",
          "INVALID_QUANTITY",
          "MARKET_CLOSED",
          "SUSPENDED",
          "KYC_DORMANT",
          "RISK_CHECK",
          "OI_LIMIT",
          "QUANTITY_LIMIT",
          "VALUE_LIMIT",
          "SEGMENT_NOT_ALLOWED",
          "ANCHOR_ACTIVE",
          "CO_APPROVAL_REQUIRED",
          "OTHER"
        ],
        "title": "RejectionReason"
      },
      "ReviewRequest": {
        "properties": {
          "order_hash": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Order Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "order_hash"
        ],
        "title": "ReviewRequest"
      },
...
      "Side": {
        "type": "string",
        "enum": [
          "BUY",
          "SELL"
        ],
        "title": "Side"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Validity": {
        "type": "string",
        "enum": [
          "DAY",
          "IOC"
        ],
        "title": "Validity"
      },
...
    }
  }
}